Add ability to write out the code into directories.
So that I can test the compiler, the reader now will write out a bunch
of files in a directory. This should let me test Norcroft against the
examples. It can also optionally write out the intro and task sections
from the page, which might be useful in understanding what's intended.