Ensure that the names produced are UTF-8.
There was some confusion in the naming of the files and the internal
`name` properties. Now it should be the case that internally we deal
with unicode names, and on the command line we supply UTF-8 names.

The command line is obviously UTF-8 in order to ensure that we can supply
all the names we need on the command line, but internally the choice
to use unicode was driven by what I feel is the more sane of the
choices - that the native python representation is used internally,
and only at presentation time does the user need to worry about how
that is encoding.

That also affects fsname, which will now be forced to UTF-8, rather
than just taking the wiki name, which might be percent encoded.